Module codec/register/loadbalance/serialize finished initially.

+++ compact-rpc-base/src/main/java/com/tsy/rpc/base/message/Message.java
@@ -13,8 +13,7 @@ public abstract class Message implements Serializable {
 
     private static final long serialVersionUID = 42L;
 
-    private long sequenceId;
+    private int sequenceId;
 
     private byte type;
-
 }
